Type casting bug in STM32HWEncoder fixed

+++ src/encoders/stm32hwencoder/STM32HWEncoder.h
@@ -16,7 +16,7 @@ class STM32HWEncoder : public Sensor {
     Encoder class constructor
     @param ppr  impulses per rotation  (cpr=ppr*4)
     */
-    explicit STM32HWEncoder(unsigned int ppr, int8_t pinA, int8_t pinB, int8_t pinI=-1);
+    explicit STM32HWEncoder(unsigned int ppr, int pinA, int pinB, int pinI=-1);
 
     void init() override;
     int needsSearch() override;
